requirements.txt
flask
twilio
gunicorn
openai>=1.0.0
google-api-python-client
google-auth
google-auth-oauthlib
